Move service deps to their own file
