+++ src/fides/api/api/deps.py
@@ -10,10 +10,6 @@ from fides.api.util.cache import get_cache as get_redis_connection
 from fides.config import CONFIG, FidesConfig
 from fides.config import get_config as get_app_config
 from fides.config.config_proxy import ConfigProxy
-from fides.service.dataset.dataset_config_service import DatasetConfigService
-from fides.service.dataset.dataset_service import DatasetService
-from fides.service.messaging.messaging_service import MessagingService
-from fides.service.privacy_request.privacy_request_service import PrivacyRequestService
 
 _engine = None
 
@@ -70,27 +66,3 @@ def get_cache() -> Generator:
             "Application redis cache required, but it is currently disabled! Please update your application configuration to enable integration with a Redis cache."
         )
     yield get_redis_connection()
-
-
-def get_messaging_service(
-    db: Session = Depends(get_db),
-    config: FidesConfig = Depends(get_config),
-    config_proxy: ConfigProxy = Depends(get_config_proxy),
-) -> MessagingService:
-    return MessagingService(db, config, config_proxy)
-
-
-def get_privacy_request_service(
-    db: Session = Depends(get_db),
-    config_proxy: ConfigProxy = Depends(get_config_proxy),
-    messaging_service: MessagingService = Depends(get_messaging_service),
-) -> PrivacyRequestService:
-    return PrivacyRequestService(db, config_proxy, messaging_service)
-
-
-def get_dataset_service(db: Session = Depends(get_db)) -> DatasetService:
-    return DatasetService(db)
-
-
-def get_dataset_config_service(db: Session = Depends(get_db)) -> DatasetConfigService:
-    return DatasetConfigService(db)
